+++ adapters/AppLovin/CHANGELOG.md
@@ -1,5 +1,6 @@
 # AppLovin Adapter for Google Mobile Ads SDK for iOS
 
+
 ## Version 5.0.1.0
 - Verified compatibility with Applovin SDK 5.0.1.